Working call for x64 (fixed test a bit). A lot of debug messages I need to remove
(cherry picked from commit 0ee46938775b24efe6fc0b858f343bbe40b46f9a)

# Conflicts:
#	angrop/chain_builder/builder.py
#	angrop/chain_builder/reg_setter.py
#	tests/test_csu_init.py
